+++ changelog.md
@@ -11,6 +11,7 @@ __[maintenance]__
 * Update certificate information for LibreOffice installers.
 * Update silent install parameters for Shotcut due to installer change from NSIS
   to Inno Setup.
+* Update certificate information for PDF24 Creator installer.
 
 ## Version 2023.09.14.0